app/requirements.txt
python-dotenv==1.0.1
langchain==0.1.13
langchain-community==0.0.31
openai==0.28
sentence-transformers==2.2.2
qdrant_client==1.8.2